backend/requirements.txt
asgiref==3.8.1
django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
sqlparse==0.4.4
tzdata==2023.3
